pyproject.toml
[tool.poetry]
name = "oz-main-project-be"
version = "0.1.0"
description = ""
authors = ["yyysolhhh <solbaram37@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
django-environ = "^0.11.2"
djangorestframework = "^3.15.1"
django-extensions = "^3.2.3"
ipython = "^8.24.0"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
gunicorn = "^22.0.0"
django-redis = "^5.4.0"
channels = "^4.1.0"
django-stubs-ext = "^5.0.0"
djangorestframework-stubs = "^3.15.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"
django-stubs = "^5.0.0"
coverage = "^7.5.0"
django-debug-toolbar = "^4.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.12
strict = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.settings"

[tool.coverage.run]
omit = ["*/migrations/*"]
